Add WebAssembly web app, refactor simulation core
C++ changes:
- Quarter-TC bucketing (65 slots, -8.0 to +8.0 in 0.25 steps)
- Surrender expanded to None/Late/Early (int 0/1/2)
- Add BJ65 rule (6:5 vs 3:2 payout toggle)
- Replace Players array with single tracked player + numOtherPlayers dummies
- Extract runShoes() core; add runBatch() for WASM single-threaded entry point
- trueCount stored as float throughout; strategies.cpp updated accordingly
- Progress output moved to stderr; JSON to stdout for native build

New files:
- wasm_api.cpp: Emscripten exported C API (bjsim_configure, bjsim_run_batch,
  bjsim_reset, bjsim_get_results)
- Makefile: native / wasm / test targets
- deploy.sh: one-shot Ubuntu 24.04 build + nginx deploy
- public/index.html, style.css, app.js, worker.js: single-page web app

Web app:
- Web Workers + independent WASM instances per thread (no SharedArrayBuffer)
- Real-time EV/RoR updates as each 500-shoe worker batch completes
- Full betting strategy table (TC threshold → hands → bet in 0.25 TC steps)
- EV/hour, house edge, Risk of Ruin, frequency and EV-by-TC charts
- Export JSON hook for future server-side result caching

+++ include/strategies.h
@@ -8,5 +8,5 @@ class strategies{
         decisions play(hand& hand, int& upcard, shoe& shoe, int strategy);
         decisions dealer(hand& hand);
         decisions playerBasic(hand& hand, int upCard);      
-        decisions playerDeviations(hand& hand, int upCard, int trueCount, int runningCount);  
+        decisions playerDeviations(hand& hand, int upCard, float trueCount, int runningCount);
 };
